+++ edison-guava/src/main/java/de/otto/edison/cachestatistics/CacheStatisticsProvider.java
@@ -1,7 +1,10 @@
 package de.otto.edison.cachestatistics;
 
+import de.otto.edison.annotations.Beta;
+
 import java.util.List;
 
+@Beta
 public interface CacheStatisticsProvider {
 
     List<CacheStatistics> getStats();