+++ HISTORY.md
@@ -4,8 +4,10 @@
 
 New Features:
 * Broken Jobs are automatically restarted according to the number of retries specified in the ```JobDefintion```.
+* ```JobDefinition``` now has a new field ```restarts```, specifying how often a job is restarted if it failed because of an error.
 
 Breaking Changes:
+* The factory methods of the ```DefaultJobDefinition``` now have an additional parameter for the number of restarts.
 * Every ```JobRunnable``` now needs to provide a ```JobDefinition```
 * Removed ```JobRunnable#getJobType```. The job type is accessed by the ```JobDefinition```